+++ serveur-express-taches/package.json
@@ -10,6 +10,8 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "@types/express": "^5.0.0",
+    "@types/nodemailer": "^6.4.16",
     "express": "^4.21.1",
     "nodemailer": "^6.9.16"
   },
