+++ serveur-express-taches/tsconfig.json
@@ -1,0 +1,16 @@
+{
+    "compilerOptions": {
+        "target": "es6",
+        "module": "commonjs",
+        "strict": true,
+        "esModuleInterop": true,
+        "forceConsistentCasingInFileNames": true,
+        "skipLibCheck": true,
+        "outDir": "./dist",
+        "rootDir": "./src",
+        "noImplicitAny": true,
+        "strictNullChecks": true
+    },
+    "include": ["./src/**/*.ts"],
+    "exclude": ["./node_modules"]
+}
